Moves utils to index.ts
